+++ packages/common/src/utils/filters.mock.ts
@@ -1,5 +1,6 @@
 import { ConditionalOperator } from '../types/conditionalRule';
-import { type Table } from '../types/explore';
+import { SupportedDbtAdapter } from '../types/dbt';
+import { type Explore, type Table } from '../types/explore';
 import {
     CustomDimensionType,
     DimensionType,
@@ -14,7 +15,8 @@ import {
     type FilterGroup,
     type FilterRule,
     type Filters,
-    type MetricFilterRule,
+    type JoinModelRequiredFilterRule,
+    type ModelRequiredFilterRule,
     type OrFilterGroup,
 } from '../types/filter';
 import type { MetricQuery } from '../types/metricQuery';
@@ -280,7 +282,9 @@ export const filterRule: FilterRule = {
     values: ['mockValue1', 'mockValue2'],
 };
 
-export const metricFilterRule = (inputFieldRef: string): MetricFilterRule => ({
+export const modelRequiredFilterRule = (
+    inputFieldRef: string,
+): ModelRequiredFilterRule => ({
     id: 'uuid',
     operator: FilterOperator.IN_THE_NEXT,
     settings: {
@@ -291,6 +295,20 @@ export const metricFilterRule = (inputFieldRef: string): MetricFilterRule => ({
     },
     values: [14],
 });
+
+export const joinedModelRequiredFilterRule = (
+    inputFieldRef: string,
+    tableName: string,
+): JoinModelRequiredFilterRule => {
+    const requiredFilterRule = modelRequiredFilterRule(inputFieldRef);
+    return {
+        ...requiredFilterRule,
+        target: {
+            ...requiredFilterRule.target,
+            tableName,
+        },
+    };
+};
 
 export const baseTable: Omit<Table, 'lineageGraph'> = {
     name: 'table',
@@ -335,6 +353,147 @@ export const expectedRequiredResetResult: FilterGroup = {
             settings: {
                 unitOfTime: 'years',
             },
+        },
+    ],
+};
+
+export const mockExplore: Explore = {
+    name: 'test',
+    label: 'Test',
+    tags: [],
+    baseTable: 'orders',
+    targetDatabase: SupportedDbtAdapter.POSTGRES,
+    joinedTables: [],
+    tables: {
+        orders: {
+            name: 'orders',
+            label: 'Orders',
+            database: 'test',
+            schema: 'public',
+            sqlTable: 'orders',
+            metrics: {},
+            lineageGraph: {},
+            dimensions: {
+                order_date_year: {
+                    name: 'order_date_year',
+                    label: 'Order Date Year',
+                    table: 'orders',
+                    tableLabel: 'Orders',
+                    compiledSql: 'order_date_year',
+                    tablesReferences: [],
+                    sql: 'order_date_year',
+                    hidden: false,
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.DATE,
+                    timeIntervalBaseDimensionName: 'order_date',
+                },
+                order_date_month: {
+                    name: 'order_date_month',
+                    label: 'Order Date Month',
+                    table: 'orders',
+                    tableLabel: 'Orders',
+                    compiledSql: 'order_date_month',
+                    tablesReferences: [],
+                    sql: 'order_date_month',
+                    hidden: false,
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.DATE,
+                    timeIntervalBaseDimensionName: 'order_date',
+                },
+                order_date_week: {
+                    name: 'order_date_week',
+                    label: 'Order Date Week',
+                    table: 'orders',
+                    tableLabel: 'Orders',
+                    compiledSql: 'order_date_week',
+                    sql: 'order_date_week',
+                    hidden: false,
+                    tablesReferences: [],
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.DATE,
+                    timeIntervalBaseDimensionName: 'order_date',
+                },
+                status: {
+                    name: 'status',
+                    label: 'Status',
+                    table: 'orders',
+                    tableLabel: 'Orders',
+                    compiledSql: 'status',
+                    sql: 'status',
+                    hidden: false,
+                    tablesReferences: [],
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.BOOLEAN,
+                },
+            },
+        },
+    },
+};
+
+export const mockExploreWithJoinedTable: Explore = {
+    ...mockExplore,
+    tables: {
+        orders: {
+            ...mockExplore.tables.orders,
+            dimensions: {
+                ...mockExplore.tables.orders.dimensions,
+                customer_id: {
+                    name: 'customer_id',
+                    label: 'Customer ID',
+                    table: 'orders',
+                    tableLabel: 'Orders',
+                    compiledSql: 'customer_id',
+                    sql: 'customer_id',
+                    hidden: false,
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.NUMBER,
+                    tablesReferences: [],
+                },
+            },
+        },
+        customers: {
+            name: 'customers',
+            label: 'Customers',
+            database: 'test',
+            schema: 'public',
+            sqlTable: 'customers',
+            metrics: {},
+            lineageGraph: {},
+            dimensions: {
+                id: {
+                    name: 'id',
+                    label: 'ID',
+                    table: 'customers',
+                    tableLabel: 'Customers',
+                    compiledSql: 'id',
+                    tablesReferences: [],
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.STRING,
+                    sql: 'id',
+                    hidden: false,
+                },
+                created_at_week: {
+                    name: 'created_at_week',
+                    label: 'Created At Week',
+                    table: 'customers',
+                    tableLabel: 'Customers',
+                    compiledSql: 'created_at_week',
+                    sql: 'created_at_week',
+                    hidden: false,
+                    tablesReferences: [],
+                    fieldType: FieldType.DIMENSION,
+                    type: DimensionType.DATE,
+                    timeIntervalBaseDimensionName: 'created_at',
+                },
+            },
+        },
+    },
+    joinedTables: [
+        {
+            table: 'customers',
+            sqlOn: '${orders.customer_id} = ${customers.id}',
+            compiledSqlOn: '(orders.customer_id) = (customers.id)',
+            type: undefined,
         },
     ],
 };